Journal Authentication

+++ src/main/java/com/journal/journalApp/service/JournalEntryService.java
@@ -27,7 +27,7 @@ public class JournalEntryService {
 			User user = userService.findByUserName(userName);
 			journalEntry saved= journalEntryRepository.save(entry);
 			user.getJournalEntries().add(saved);
-			userService.saveEntry(user);
+			userService.saveUser(user);
 		}
 		catch(Exception e) {
 			System.out.println(e);
@@ -49,10 +49,23 @@ public class JournalEntryService {
 		return journalEntryRepository.findById(id);
 	}
 	
-	public void deleteById(ObjectId id, String userName) {
+	@Transactional
+	public boolean deleteById(ObjectId id, String userName) {
+		boolean removed = false;
+		try {
 		User user = userService.findByUserName(userName);
-		user.getJournalEntries().removeIf(x->x.getId().equals(id));
-		userService.saveEntry(user);
+		removed= user.getJournalEntries().removeIf(x->x.getId().equals(id));
+		if(removed){
+		userService.saveUser(user);
 		journalEntryRepository.deleteById(id);
+		}
+		}
+		catch(Exception e) {
+			System.out.println(e);
+			throw new RuntimeException("An error occurred while deleting the entry",e);
+		}
+		return removed;
 	}
+	
+	
 }
